Finished implementation of ducks needed UI
Treated each bars worth of duck like an object containing 4 bars in a container with layout group to handle scaling.

These spawn on start and set the groups image children alphas to 0 of the ducks not needed, these groups incrementally have their groups set to 1 alpha when they appear and flash to better notify the player!