Added pop up screen space UI for score
Added a new canvas which is just for the pop up of the score when a player shoots a duck.

This also lead me to fix a bug which was all ducks in the pool having their on notify invoked. Simple fix of adding a check that the instance is active before invoking and observer behaviors'