Refactored unity events to Interface based ovserver
Refactored the UI Event list in round manager by creating round subject, I would say there are more calls in the refactored method, but invokes are less obscure to the array[i].invoke (ms is also the same as before, so impact is negligible). Also allows values to be passed and defined in subject interface as needed.

With the round state being used as a flag to determine whether the player can shoot or not!